backend/requirements.txt
appnope==0.1.0
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
certifi==2017.7.27.1
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
decorator==4.1.2
Django==1.11.3
django-cors-headers==2.1.0
django-environ==0.4.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
factory-boy==2.9.0
Faker==0.7.18
fancycompleter==0.7
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
mock==2.0.0
msgpack-python==0.4.8
packaging==16.8
pbr==3.1.1
pdbpp==0.9.1
pexpect==4.2.1
pickleshare==0.7.4
pip-review==0.5.3
prometheus-client==0.0.20
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
PyJWT==1.5.2
pyparsing==2.2.0
pytest==3.1.3
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.5
requests==2.18.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
Twisted==17.5.0
txaio==2.8.1
urllib3==1.22
uWSGI==2.0.15
wcwidth==0.1.7
wmctrl==0.3
zope.interface==4.4.2
